analytics-docker/requirements.txt
streamlit
snowflake-connector-python
snowflake
pandas
numpy
requests
datetime
yfinance
python-dotenv
matplotlib
